+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 
+### Fixed
+
+- Module name will now be included in hover and signature displays for imported types
+
 ## [1.5.0] - 2022-06-20
 
 ### Added